added new ressources + firebase files and gradle conf
