+++ app/build.gradle
@@ -1,13 +1,14 @@
 apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-android-extensions'
+apply plugin: 'com.google.gms.google-services'
 
 android {
     compileSdkVersion 30
     buildToolsVersion "30.0.0"
 
     defaultConfig {
-        applicationId "com.example.lokakaka"
+        applicationId "com.fontbenjamin.lokakaka"
         minSdkVersion 16
         targetSdkVersion 30
         versionCode 1
@@ -25,6 +26,7 @@ android {
 }
 
 dependencies {
+    implementation platform('com.google.firebase:firebase-bom:25.12.0')
     implementation fileTree(dir: "libs", include: ["*.jar"])
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
     implementation 'androidx.core:core-ktx:1.3.1'
