--- app/src/main/java/com/example/lokakaka/SplashScreenActivity.kt
@@ -1,13 +1,0 @@
-package com.example.lokakaka
-
-import android.content.Intent
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-
-class SplashScreenActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        startActivity(Intent(this@SplashScreenActivity, MainActivity::class.java))
-        finish()
-    }
-}
