+++ app/src/main/java/com/fontbenjamin/lokakaka/SplashScreenActivity.kt
@@ -1,0 +1,13 @@
+package com.fontbenjamin.lokakaka
+
+import android.content.Intent
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+
+class SplashScreenActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        startActivity(Intent(this@SplashScreenActivity, MainActivity::class.java))
+        finish()
+    }
+}
