+++ app/src/main/java/com/fontbenjamin/lokakaka/chatFragment/ChatFragment.kt
@@ -1,0 +1,19 @@
+package com.fontbenjamin.lokakaka.chatFragment
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.fontbenjamin.lokakaka.R
+
+class ChatFragment: Fragment() {
+
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? =
+        inflater.inflate(R.layout.fragment_chat, container, false)
+
+    companion object {
+        fun newInstance(): ChatFragment = ChatFragment()
+    }
+
+}
