+++ app/src/main/res/values/strings.xml
@@ -3,12 +3,12 @@
     <string name="menu_welcome">Salutations</string>
     <string name="menu_encyclopedia">Activités</string>
     <string name="menu_trading">Routes</string>
-    <string name="menu_chat">Discord</string>
+    <string name="menu_chat">Contrats</string>
     <string name="title_full">Lokakaka!Org█</string>
     <string name="title">Lokakaka!Org</string>
 
     <string name="credits">Credits</string>
-    <string name="credits_content">Application réalisée par PrincessOctopus\n\nCalculateur de routes commerciales réalisé par Kamille92\n\nIcones trouvées sur www.flaticon.com :\n\ndragon-fruit / drug / pirate-skull / checkered-flag / binary-code / saber / circuit / target / pirate-hat / treasure-chest / chat-box / skull / mouse / marriage-certificate : Icones réalisées par Freepik\nrocket : Icone réalisée par fjstudio\nbox : Icone réalisée par Gregor Cresnar\nyoutube / discord : Icones réalisées par Pixel perfect\nmenu: Icone réalisée par Those Icons\ngarbage : Icone réalisée par Nhor Phai\nbackarrow : Icone réalisée par Kiranshastry</string>
+    <string name="credits_content">Calculateur de routes commerciales réalisé par Kamille92\n\nIcones trouvées sur www.flaticon.com :\n\ndragon-fruit / drug / pirate-skull / checkered-flag / binary-code / saber / circuit / target / pirate-hat / treasure-chest / chat-box / skull / mouse / marriage-certificate : Icones réalisées par Freepik\nrocket : Icone réalisée par fjstudio\nbox : Icone réalisée par Gregor Cresnar\nyoutube / discord : Icones réalisées par Pixel perfect\nmenu: Icone réalisée par Those Icons\ngarbage : Icone réalisée par Nhor Phai\nbackarrow : Icone réalisée par Kiranshastry</string>
     <string name="welcome_text">Si vous êtes intéressé par la contrebande, le piratage, ou toute autre activité illégale qui peut vous donner accès à beaucoup d\’UECs, alors nous avons quelque chose en commun ! Pourquoi ne pas rejoindre notre organisation ?\n\nDans ce syndicat, l\’objectif n’est pas de vous forcer à jouer d’une façon ou d’une autre. Les joueurs occasionnels sont les bienvenus de même que les pirates confirmés, et l’ambiance générale est que nous sommes ici pour passer un bon moment ensemble dans les recoins obscurs du Verse, tout en nous entraidans les uns les autres. Donc, si tu veux partager ta sagesse criminelle, apprendre à opérer en dehors de la loi, ou si tu veux juste rencontrer des gens pour apporter un Caterpillar plein à craquer de Neon à bon port, soit le bienvenu dans la famille Lokakaka.</string>
     <string name="button_syndicate">La page du syndicat</string>
     <string name="button_youtube">La chaine Youtube</string>
